Added Invite Friends Page in models
The page was added in the home page models. The page
would later be used to send out friend invitations
in order to register with the social media website.

To do the same the Invite Friends page should be
created from the admin and be made a child of
the home page for the link to get reflected on the
home page.